A bit more overflow prevention for high step rates

+++ Due/Marlin/stepper.cpp
@@ -61,10 +61,10 @@ volatile static unsigned long step_events_completed; // The number of step event
 #endif
 static long acceleration_time, deceleration_time;
 //static unsigned long accelerate_until, decelerate_after, acceleration_rate, initial_rate, final_rate, nominal_rate;
-static unsigned short acc_step_rate; // needed for deccelaration start point
+static unsigned long acc_step_rate; // needed for deccelaration start point
 static char step_loops;
 static unsigned long OCR1A_nominal;
-static unsigned short step_loops_nominal;
+static unsigned long step_loops_nominal;
 
 volatile long endstops_trigsteps[3]={0,0,0};
 volatile long endstops_stepsTotal,endstops_stepsDone;
@@ -258,10 +258,7 @@ void step_wait(){
 
 
 FORCE_INLINE unsigned long calc_timer(unsigned long step_rate) {
-  
- //  SERIAL_ECHOPGM("calc timer: ");
- //  SERIAL_ECHOLN(step_rate);
-  
+    
   unsigned long timer;
   if(step_rate > MAX_STEP_FREQUENCY) step_rate = MAX_STEP_FREQUENCY;
